+++ Godeps/_workspace/src/github.com/emicklei/go-restful/swagger/swagger_webservice.go
@@ -5,11 +5,12 @@ import (
 
 	"github.com/emicklei/go-restful"
 	// "github.com/emicklei/hopwatch"
-	"log"
 	"net/http"
 	"reflect"
 	"sort"
 	"strings"
+
+	"github.com/emicklei/go-restful/log"
 )
 
 type SwaggerService struct {
@@ -24,7 +25,10 @@ func newSwaggerService(config Config) *SwaggerService {
 }
 
 // LogInfo is the function that is called when this package needs to log. It defaults to log.Printf
-var LogInfo = log.Printf
+var LogInfo = func(format string, v ...interface{}) {
+	// use the restful package-wide logger
+	log.Printf(format, v...)
+}
 
 // InstallSwaggerService add the WebService that provides the API documentation of all services
 // conform the Swagger documentation specifcation. (https://github.com/wordnik/swagger-core/wiki).
@@ -71,6 +75,11 @@ func RegisterSwaggerService(config Config, wsContainer *restful.Container) {
 				sws.apiDeclarationMap[each.RootPath()] = sws.composeDeclaration(each, each.RootPath())
 			}
 		}
+	}
+
+	// if specified then call the PostBuilderHandler
+	if config.PostBuildHandler != nil {
+		config.PostBuildHandler(sws.apiDeclarationMap)
 	}
 
 	// Check paths for UI serving