+++ Godeps/_workspace/src/github.com/emicklei/go-restful/web_service_test.go
@@ -85,6 +85,50 @@ func TestSelectedRoutePath_Issue100(t *testing.T) {
 	}
 }
 
+func TestContentType415_Issue170(t *testing.T) {
+	tearDown()
+	Add(newGetOnlyJsonOnlyService())
+	httpRequest, _ := http.NewRequest("GET", "http://here.com/get", nil)
+	httpWriter := httptest.NewRecorder()
+	DefaultContainer.dispatch(httpWriter, httpRequest)
+	if 200 != httpWriter.Code {
+		t.Errorf("Expected 200, got %d", httpWriter.Code)
+	}
+}
+
+func TestContentType415_POST_Issue170(t *testing.T) {
+	tearDown()
+	Add(newPostOnlyJsonOnlyService())
+	httpRequest, _ := http.NewRequest("POST", "http://here.com/post", nil)
+	httpRequest.Header.Set("Content-Type", "application/json")
+	httpWriter := httptest.NewRecorder()
+	DefaultContainer.dispatch(httpWriter, httpRequest)
+	if 200 != httpWriter.Code {
+		t.Errorf("Expected 200, got %d", httpWriter.Code)
+	}
+}
+
+// go test -v -test.run TestContentTypeOctet_Issue170 ...restful
+func TestContentTypeOctet_Issue170(t *testing.T) {
+	tearDown()
+	Add(newGetConsumingOctetStreamService())
+	// with content-type
+	httpRequest, _ := http.NewRequest("GET", "http://here.com/get", nil)
+	httpRequest.Header.Set("Content-Type", MIME_OCTET)
+	httpWriter := httptest.NewRecorder()
+	DefaultContainer.dispatch(httpWriter, httpRequest)
+	if 200 != httpWriter.Code {
+		t.Errorf("Expected 200, got %d", httpWriter.Code)
+	}
+	// without content-type
+	httpRequest, _ = http.NewRequest("GET", "http://here.com/get", nil)
+	httpWriter = httptest.NewRecorder()
+	DefaultContainer.dispatch(httpWriter, httpRequest)
+	if 200 != httpWriter.Code {
+		t.Errorf("Expected 200, got %d", httpWriter.Code)
+	}
+}
+
 func newPanicingService() *WebService {
 	ws := new(WebService).Path("")
 	ws.Route(ws.GET("/fire").To(doPanic))
@@ -94,6 +138,27 @@ func newPanicingService() *WebService {
 func newGetOnlyService() *WebService {
 	ws := new(WebService).Path("")
 	ws.Route(ws.GET("/get").To(doPanic))
+	return ws
+}
+
+func newPostOnlyJsonOnlyService() *WebService {
+	ws := new(WebService).Path("")
+	ws.Consumes("application/json")
+	ws.Route(ws.POST("/post").To(doNothing))
+	return ws
+}
+
+func newGetOnlyJsonOnlyService() *WebService {
+	ws := new(WebService).Path("")
+	ws.Consumes("application/json")
+	ws.Route(ws.GET("/get").To(doNothing))
+	return ws
+}
+
+func newGetConsumingOctetStreamService() *WebService {
+	ws := new(WebService).Path("")
+	ws.Consumes("application/octet-stream")
+	ws.Route(ws.GET("/get").To(doNothing))
 	return ws
 }
 
@@ -112,4 +177,7 @@ func selectedRouteChecker(req *Request, resp *Response) {
 func doPanic(req *Request, resp *Response) {
 	println("lightning...")
 	panic("fire")
+}
+
+func doNothing(req *Request, resp *Response) {
 }